// proyecto-final/src/app/nasa/interfaces/mars-rover-interface.ts
export interface MarsRoverResponse {
  photos: MarsPhoto[];
}

export interface MarsPhoto {
  id: number;
  sol: number;
  camera: Camera;
  img_src: string;
  earth_date: string;
  rover: Rover;
}

export interface Camera {
  id: number;
  name: string;
  rover_id: number;
  full_name: string;
}

export interface Rover {
  id: number;
  name: string;
  landing_date: string;
  launch_date: string;
  status: string;
}